requirements.txt
# Core Trading Dependencies
alpaca-py>=0.43.0                    # Alpaca API client (premium data support)
alpaca-trade-api>=3.2.0              # Legacy API for additional features

# Data Processing & Analysis
pandas>=2.0.0                         # Data manipulation
numpy>=1.24.0                         # Numerical computing
TA-Lib>=0.4.28                        # Technical analysis indicators (requires binary install)
pandas-ta>=0.3.14b                    # Additional technical indicators (TA-Lib alternative)
scipy>=1.10.0                         # Statistical functions

# Database
sqlalchemy==2.0.25                    # ORM and database toolkit
psycopg2-binary==2.9.9               # PostgreSQL adapter (optional, use SQLite initially)
alembic==1.13.1                       # Database migrations

# Machine Learning & AI
scikit-learn==1.4.0                   # ML algorithms for pattern recognition
anthropic==0.18.1                     # Claude AI API
openai==1.12.0                        # OpenAI GPT API (alternative)
langchain==0.1.6                      # LLM orchestration
langchain-community==0.0.20           # Community integrations

# Time & Scheduling
pytz==2024.1                          # Timezone handling
schedule==1.2.1                       # Job scheduling
APScheduler==3.10.4                   # Advanced Python scheduler

# Utilities
python-dotenv==1.0.1                  # Environment variable management
pydantic==2.6.0                       # Data validation
pydantic-settings==2.1.0              # Settings management
python-json-logger==2.0.7             # Structured JSON logging
retry==0.9.2                          # Retry decorator
tenacity==8.2.3                       # Advanced retry logic

# Monitoring & Logging
loguru==0.7.2                         # Advanced logging
prometheus-client==0.20.0             # Metrics collection

# Configuration & Testing
pytest==8.0.0                         # Testing framework
pytest-asyncio==0.23.4                # Async test support
pytest-mock==3.12.0                   # Mocking support
python-dateutil==2.8.2                # Date utilities

# Performance
# numba==0.59.0                       # JIT compilation for speed (incompatible with Python 3.11+)
joblib>=1.3.0                         # Parallel processing

# API & Networking
requests==2.31.0                      # HTTP library
websockets==12.0                      # WebSocket client
aiohttp==3.9.3                        # Async HTTP

# Backtesting
vectorbt==0.26.1                      # Vectorized backtesting (optional, heavy dependency)
backtrader==1.9.78.123                # Alternative backtesting engine
